apps/Gemini_Voice/requirements.txt
google-genai
python-dotenv
pydub
fastapi
uvicorn
edge-tts
python-multipart
